+++ .rebase/CHANGELOG.md
@@ -2,6 +2,41 @@
 
 The file to keep a list of changed files which will potentionaly help to resolve rebase conflicts.
 
+#### @rnikitenko
+https://github.com/che-incubator/che-code/pull/735
+
+- code/extensions/copilot/.esbuild.ts
+---
+
+#### @RomanNikitenko
+https://github.com/che-incubator/che-code/pull/736
+
+- code/extensions/git/src/askpass.sh
+- code/extensions/git/src/ssh-askpass.sh
+- code/extensions/git/src/git-editor.sh
+- code/extensions/copilot/src/extension/chatSessions/vscode-node/copilotCLITerminalIntegration.ts
+- code/extensions/copilot/src/extension/onboardDebug/vscode-node/copilotDebugCommandContribution.ts
+---
+
+#### @sbouchet
+https://github.com/che-incubator/che-code/pull/739
+
+- code/package.json
+- code/remote/package.json
+- code/extensions/copilot/package.json
+- code/extensions/copilot/chat-lib/package.json
+- code/extensions/notebook-renderers/package.json
+---
+
+#### @sbouchet
+https://github.com/che-incubator/che-code/pull/728
+
+- code/package.json
+- code/remote/package.json
+- code/build/rspack/package.json
+- code/extensions/copilot/package.json
+---
+
 #### @sbouchet
 https://github.com/che-incubator/che-code/pull/717
 
@@ -26,6 +61,16 @@ https://github.com/che-incubator/che-code/pull/720
 ---
 
 #### @sbouchet
+https://github.com/che-incubator/che-code/pull/722
+
+- code/package.json
+- code/remote/package.json
+- code/build/npm/gyp/package.json
+- code/extensions/copilot/package.json
+- code/test/mcp/package.json
+---
+
+#### @sbouchet
 https://github.com/che-incubator/che-code/pull/716
 
 - code/package.json
@@ -42,12 +87,14 @@ https://github.com/che-incubator/che-code/pull/705
 - code/src/vs/base/browser/dompurify/cgmanifest.json
 - code/src/vs/base/browser/dompurify/dompurify.d.ts
 - code/src/vs/base/browser/dompurify/dompurify.js
+---
 
 #### @sbouchet
 https://github.com/che-incubator/che-code/pull/711
 
 - code/package.json
 - code/remote/package.json
+---
 
 #### @sbouchet
 https://github.com/che-incubator/che-code/pull/708